+++ src/pydidas/widgets/misc/show_integration_roi_params_widget.py
@@ -1,6 +1,6 @@
 # This file is part of pydidas
 #
-# Copyright 2023 - 2025, Helmholtz-Zentrum Hereon
+# Copyright 2023 - 2026, Helmholtz-Zentrum Hereon
 # SPDX-License-Identifier: GPL-3.0-only
 #
 # pydidas is free software: you can redistribute it and/or modify
@@ -21,7 +21,7 @@ integration region of an associated Plugin in the linked plot.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2023 - 2025, Helmholtz-Zentrum Hereon"
+__copyright__ = "Copyright 2023 - 2026, Helmholtz-Zentrum Hereon"
 __license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Production"
@@ -34,7 +34,7 @@ from qtpy import QtCore
 
 from pydidas.core import get_generic_param_collection
 from pydidas.core.constants import (
-    FONT_METRIC_PARAM_EDIT_WIDTH,
+    FONT_METRIC_CONFIG_WIDTH,
     POLICY_FIX_EXP,
     POLICY_MIN_MIN,
 )
@@ -111,7 +111,7 @@ class ShowIntegrationRoiParamsWidget(WidgetWithParameterCollection):
                 "angular direction is clockwise. \nThe zero position is at the positive"
                 " x-axis, i.e. right of the beamcenter.",
                 font_metric_height_factor=7,
-                font_metric_width_factor=FONT_METRIC_PARAM_EDIT_WIDTH,
+                font_metric_width_factor=FONT_METRIC_CONFIG_WIDTH,
                 parent_widget=self._widgets["plugin_container"],
                 sizePolicy=POLICY_FIX_EXP,
                 wordWrap=True,